+++ src/model.rs
@@ -4,6 +4,7 @@ use std::path::Path;
 use std::process::Command;
 use std::str::FromStr;
 
+//use anstyle::{AnsiColor, Effects, Style};
 use chrono::{NaiveDate, Utc};
 use clap::ValueEnum;
 use indexmap::IndexMap;
@@ -215,10 +216,50 @@ pub struct Config {
     pub priority_default: Priority,
 }
 
+#[derive(Debug, Deserialize, Clone, Copy)]
+#[serde(rename_all = "snake_case")]
+pub enum NamedColor {
+    Black,
+    BrightBlack,
+    Red,
+    BrightRed,
+    Green,
+    BrightGreen,
+    Yellow,
+    BrightYellow,
+    Blue,
+    BrightBlue,
+    Magenta,
+    BrightMagenta,
+    Cyan,
+    BrightCyan,
+    White,
+    BrightWhite,
+    Bold,
+}
+
+#[derive(Debug, Deserialize, Clone)]
+pub struct Colors {
+    pub header: NamedColor,
+    pub me: NamedColor,
+    pub due_date_overdue: NamedColor,
+    pub state: IndexMap<String, NamedColor>,
+    pub priority: IndexMap<String, NamedColor>,
+    #[serde(rename = "type")]
+    pub type_: IndexMap<String, NamedColor>,
+}
+
+#[derive(Debug, Deserialize, Clone)]
+pub struct ListFormatting {
+    pub header_separator: bool,
+    pub colors: Colors,
+}
+
 #[derive(Debug, Deserialize)]
 pub struct Settings {
     pub editor: String,
     pub user: String,
+    pub list_formatting: ListFormatting,
 }
 
 #[derive(Debug, Deserialize)]
@@ -241,7 +282,7 @@ pub fn load_users() -> Result<Users, String> {
 
     let users: Users = match serde_yaml::from_str(&users_raw) {
         Ok(m) => m,
-        Err(_) => return Err("users.yaml malformatted.".to_string()),
+        Err(e) => return Err(format!("users.yaml malformatted: {e}")),
     };
 
     Ok(users)
@@ -258,32 +299,30 @@ pub fn is_valid_iso_date(s: &str) -> Result<bool, String> {
 }
 
 /// Validate if a state is in the list of valid states from config.
-pub fn is_valid_state(s: &str) -> Result<bool, String> {
-    let config = load_config()?;
-    Ok(config.states.contains(&s.to_string()))
+pub fn is_valid_state(config: &Config, s: &str) -> bool {
+    config.states.contains(&s.to_string())
 }
 
 /// Validate if a type is in the list of valid types from config or empty.
-pub fn is_valid_type(s: &str) -> Result<bool, String> {
-    let config = load_config()?;
-    Ok(s.is_empty() || config.types.contains(&s.to_string()))
+pub fn is_valid_type(config: &Config, s: &str) -> bool {
+    s.is_empty() || config.types.contains(&s.to_string())
 }
 
 /// Validate if an user is in the list of valid users:id from users.yaml.
-pub fn is_valid_user(s: &str) -> Result<bool, String> {
-    let users = load_users()?;
-    Ok(s.is_empty() || s == "me" || users.users.iter().any(|u| u.id == s))
+pub fn is_valid_user(users: &Users, s: &str) -> bool {
+    s.is_empty() || s == "me" || users.users.iter().any(|u| u.id == s)
 }
 
-pub fn user_handle_me(value: &mut String) -> Result<(), String> {
-    if *value == "me" {
-        let settings = load_settings()?;
-        *value = match is_valid_user(&settings.user) {
-            Ok(true) => settings.user,
-            Ok(false) => return Err("Invalid user: settings.yaml::user must be part of users.yaml:users:id or ''".to_string()),
-            Err(e) => return Err(format!("Config error: {e}")),
-        }
+pub fn user_handle_me(users: &Users, settings: &Settings, value: &mut String) -> Result<(), String> {
+    if *value != "me" {
+        return Ok(());
     }
+
+    *value = if is_valid_user(users, &settings.user) {
+        settings.user.clone()
+    } else {
+        return Err("Invalid user: settings.yaml::user must be part of users.yaml:users:id or ''".to_string());
+    };
 
     Ok(())
 }
@@ -297,7 +336,7 @@ pub fn load_config() -> Result<Config, String> {
 
     let config: Config = match serde_yaml::from_str(&config_raw) {
         Ok(m) => m,
-        Err(_) => return Err("config.yaml malformatted.".to_string()),
+        Err(e) => return Err(format!("config.yaml malformatted: {e}")),
     };
 
     Ok(config)
@@ -312,7 +351,7 @@ pub fn load_settings() -> Result<Settings, String> {
 
     let settings: Settings = match serde_yaml::from_str(&settings_raw) {
         Ok(m) => m,
-        Err(_) => return Err("settings.yaml malformatted.".to_string()),
+        Err(e) => return Err(format!("settings.yaml malformatted: {e}")),
     };
 
     Ok(settings)
@@ -326,7 +365,7 @@ pub fn load_meta(path: &Path) -> Result<Meta, String> {
 
     let meta: Meta = match serde_yaml::from_str(&meta_raw) {
         Ok(m) => m,
-        Err(_) => return Err(format!("meta.yaml malformatted: {}", path.display())),
+        Err(e) => return Err(format!("meta.yaml malformatted: {}: {e}", path.display())),
     };
 
     Ok(meta)